requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.0.10
pypdf==3.17.4
faiss-cpu==1.7.4
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
gunicorn==21.2.0
beautifulsoup4==4.12.2
PyJWT==2.8.0